corrections int->long
(Logical change 1.1161)


git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@11270 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ pm2/pm2/marcel/include/asm-generic/linux_bitops.h
@@ -98,7 +98,7 @@ static __inline__ int ma_constant_test_bit(int nr, const volatile unsigned long 
 }
 static __inline__ int ma_variable_test_bit(int nr, const unsigned long * addr)
 {
-	int	mask;
+	unsigned long	mask;
 
 	addr += nr / MA_BITS_PER_LONG;
 	mask = 1UL << (nr % MA_BITS_PER_LONG);
